+++ client/src/components/Home/CategoryContainer.tsx
@@ -2,9 +2,9 @@ import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import axios from 'axios';
 
-import CategoryCard from '../../components/common/CategoryCard';
+import CategoryCard from '../common/CategoryCard';
 import Title from './Title';
-import CreateChannelButton from '../../components/ui/CreateChannelButton';
+import CreateChannelButton from '../ui/CreateChannelButton';
 
 import iconData from '../../data/categoryIcons';
 import { CategoryType } from '../../types/dataTypes';
