+++ client/src/components/Posting/ImagePreview.tsx
