+++ client/src/components/Posting/ImageSection.tsx
