+++ client/src/components/common/CategoryTag.tsx
@@ -2,15 +2,10 @@ import styled, { css } from 'styled-components';
 import { type CategoryType } from '../../types/dataTypes';
 import { type StyledTagPropsType } from '../../types/propsTypes';
 
-const CategoryTag = ({ categoryId, categoryName }: CategoryType)  => {
-  
+const CategoryTag = ({ categoryId, categoryName }: CategoryType) => {
   const mainColor = Math.round((categoryId - 1) % 4);
 
-  return (
-    <StyledTag styleId={mainColor} >
-      {categoryName}
-    </StyledTag>
-  );
+  return <StyledTag styleId={mainColor}>{categoryName}</StyledTag>;
 };
 
 export default CategoryTag;
@@ -21,10 +16,9 @@ const StyledTag = styled.span<StyledTagPropsType>`
   padding: 5px 10px;
   color: #3e3948;
   border-radius: 5px;
-  color:  ${({ styleId }) => 
-      css`var(--category-tag-color-${styleId}, var(--category-tag-color-default));
-      `
-    };
+  color: ${({ styleId }) =>
+    css`var(--category-tag-color-${styleId}, var(--category-tag-color-default));
+      `};
   background-color: ${({ styleId }) =>
     css`var(--category-tag-bg-${styleId}, var(--category-tag-bg-default))`};
   word-break: keep-all;
